requirements.txt
python-telegram-bot
arxiv
openai
python-dotenv
psutil
apscheduler
